+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 
+## [0.4.1 - 2021-05-20]
+
+- Fixed task awaiter in the EnterspeedIngestService
+
 ## [0.4.0 - 2021-03-17]
 
 - Multi-targeted the SDK for **.NET Standard 1.1** and **.NET Standard 2.0**
 - Added IJsonSerializer interface
 - Added SystemTextJsonSerializer for .NET Standard 2.0
-- Fixed Ingest endpoint path 
+- Fixed Ingest endpoint path
   - Added version number
   - Removed /api/